#1, fixing warnings from building under 64-bit

+++ CuTest.c
@@ -11,7 +11,7 @@
  * CuStr
  *-------------------------------------------------------------------------*/
 
-char* CuStrAlloc(int size)
+char* CuStrAlloc(size_t size)
 {
 	char* newStr = (char*) malloc( sizeof(char) * (size) );
 	return newStr;
@@ -19,7 +19,7 @@ char* CuStrAlloc(int size)
 
 char* CuStrCopy(const char* old)
 {
-	int len = strlen(old);
+	size_t len = strlen(old);
 	char* newStr = CuStrAlloc(len + 1);
 	strcpy(newStr, old);
 	return newStr;
@@ -55,7 +55,7 @@ void CuStringDelete(CuString *str)
         free(str);
 }
 
-void CuStringResize(CuString* str, int newSize)
+void CuStringResize(CuString* str, size_t newSize)
 {
 	str->buffer = (char*) realloc(str->buffer, sizeof(char) * newSize);
 	str->size = newSize;
@@ -63,7 +63,7 @@ void CuStringResize(CuString* str, int newSize)
 
 void CuStringAppend(CuString* str, const char* text)
 {
-	int length;
+	size_t length;
 
 	if (text == NULL) {
 		text = "NULL";
@@ -94,9 +94,9 @@ void CuStringAppendFormat(CuString* str, const char* format, ...)
 	CuStringAppend(str, buf);
 }
 
-void CuStringInsert(CuString* str, const char* text, int pos)
+void CuStringInsert(CuString* str, const char* text, size_t pos)
 {
-	int length = strlen(text);
+	size_t length = strlen(text);
 	if (pos > str->length)
 		pos = str->length;
 	if (str->length + length + 1 >= str->size)
